Cambios de short a int

+++ src/MemorySimulator/Tasks/Task.java
@@ -1,10 +1,10 @@
 package MemorySimulator.Tasks;
 public class Task {
     private final String nombre;
-    private final short tamano;
+    private final int tamano;
     private final int tiempo;
 
-    public Task(String name, short tam, int time){
+    public Task(String name, int tam, int time){
         this.nombre = name;
         this.tamano = tam;
         this.tiempo = time;
@@ -12,7 +12,7 @@ public class Task {
     public String getName(){
         return this.nombre;
     }
-    public short getTamano(){
+    public int getTamano(){
         return this.tamano;
     }
     public int getTiempo(){
